QB-16-feat(tickets-by-company):Splitting the endpoint to align with the repository, controller, and service folders
